Custom dataset class that works with torch loader
dataset.py contains a custom Dataset class "GRAND_DC2_TracesDataset". It has a "pad" transform that zero pad the missing traces (of the untriggered antennas) and similarly pads (with -1) indices of missing antennas (out of 300).

example.ipynb uses it with a pytorch dataloader. Note that the padding is important for the loader, so that tensors have constant shapes throughout events.