UNetRes class seems to be (at least functionally) working, in the sense that it returns a tensor of the right shape.
Now it needs to be trained within a denoiser.
Note that this is the basic ResBlock based UNet,
so that the noise map has to be concatenated in
to the input tensor each time.

Note that input and output channel numbers can
be adjusted with the in_nc and out_nc parameters.

Also note that the usual UNet class, with skip
connections, only works for now wiht out_nc=in_nc,
because the output is simply added to the input as
part of the last skip connection.
To be more general, we would need to add a 1x1
convolution to adjust the shape in the skip
connection.